+++ DataBaseManager.cpp
@@ -12,11 +12,6 @@ int DataBaseManager::getLastId() const
 	return lastId;
 }
 
-void DataBaseManager::setLastId(int lastId) 
-{
-	this->lastId= lastId;
-}
-
 //Special 
 
 int DataBaseManager::toInt(const string& inputObject)
@@ -26,14 +21,15 @@ int DataBaseManager::toInt(const string& inputObject)
 	isString >> intObject;
 	return intObject;
 }
-void DataBaseManager::recordObjectBD(const ObjectBD& obj, int lastId) 
+
+void DataBaseManager::recordObjectBD(const ObjectBD& obj) 
 {
 	ofstream file("product_data.txt", ios::app);
 	if (!file) throw runtime_error("Не удалось открыть файл");
 	file << obj.toText();
 	file.close();
-	setLastId(lastId);
 }
+
 double DataBaseManager::toDouble(const string& inputObject)
 {
 	string sanitized = inputObject;
@@ -43,6 +39,7 @@ double DataBaseManager::toDouble(const string& inputObject)
 	isString >> doubleObject;
 	return doubleObject;
 }
+
 ObjectBD DataBaseManager::parseObject(const vector<string>& lines)
 {
 	ObjectBD obj;
@@ -63,19 +60,16 @@ ObjectBD DataBaseManager::parseObject(const vector<string>& lines)
 	}
 	return obj;
 }	
-ifstream DataBaseManager::openFile(const std::string& fileName)
+
+vector <ObjectBD> DataBaseManager::readObjectBD()
 {
-	ifstream file(fileName);
-	if (!file.is_open()) {
-		throw runtime_error("Не удалось открыть файл");
-	}
-	return file;
-}
-vector <ObjectBD> DataBaseManager::parseFile(ifstream& file)
-{
+	ifstream file("product_data.txt");
+	if (!file) cout << ("Не удалось открыть файл") << endl;
+	else if (file) cout << ("Удалось открыть файл") << endl;
 	vector <ObjectBD> result;
 	vector <string> buffer;
 	string line;
+
 	while (getline(file, line))
 	{
 		if (line == "===")
@@ -83,24 +77,11 @@ vector <ObjectBD> DataBaseManager::parseFile(ifstream& file)
 			ObjectBD obj = parseObject(buffer);
 			result.push_back(obj);
 			buffer.clear();
+
+			cout << "Объект добавлен!" << endl;
+			/*obj.print();*/
 		}
 		else buffer.push_back(line);
 	}
 	return result;
-}
-
-vector <ObjectBD> DataBaseManager::readObjectBD()
-{
-	try 
-	{
-		ifstream file = openFile("data/product_data.txt");
-		vector <ObjectBD> result = parseFile(file);
-		file.close();
-		return result;
-	}
-	catch (const exception& e) 
-	{
-		cout << e.what() << endl;
-		return {};
-	}
 }
