+++ product_data.txt
@@ -1,35 +1,28 @@
 id: 1
-Имя: 123
-Количество: 123
-Цена: 123,000000
-Дата регистрации: 123
-Кто зарегистрировал: 123
-===
-id: 1
-Имя: qwe
-Количество: 2
-Цена: 7,000000
-Дата регистрации: qw
-Кто зарегистрировал: e
+Имя: Moloko
+Количество: 5
+Цена: 10,000000
+Дата регистрации: 18-09-2025
+Кто зарегистрировал: Gendos
 ===
 id: 2
-Имя: wert
-Количество: 123
-Цена: 76,000000
-Дата регистрации: qwe
-Кто зарегистрировал: qwe
+Имя: Hleb
+Количество: 7
+Цена: 7,000000
+Дата регистрации: 18-09-2025
+Кто зарегистрировал: Alex
 ===
 id: 1
-Имя: 1232434
-Количество: 34
-Цена: 4543,000000
-Дата регистрации: 432534
-Кто зарегистрировал: 1432534
+Имя: 2
+Количество: 3
+Цена: 4,000000
+Дата регистрации: 5
+Кто зарегистрировал: 6
 ===
 id: 1
-Имя: 1233
-Количество: 123
-Цена: 123,000000
-Дата регистрации: 123
-Кто зарегистрировал: 123
+Имя: 2
+Количество: 3
+Цена: 4,000000
+Дата регистрации: 3
+Кто зарегистрировал: 2
 ===
